Added test to test_nifti_spheres_masker.py
